// src/schemas/userSchema.ts
import * as  z from 'zod';

export const userSchema = z.object({
    name: z.string().min(1, "Name is required"),
    email: z.string().email("Invalid email address"),
    password: z.string().min(8, "Password must be at least 8 characters"),
  });
  
  export type UserFormData = z.infer<typeof userSchema>;